pyproject.toml
[tool.poetry]
name = "python_sokoban"
version = "0.1.0"
description = ""
authors = ["Eric Harm <ericharm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
toml = "^0.10.2"
mypy = "^1.0.1"
isort = "^5.12.0"
attrs = "^22.2.0"
black = "^23.1.0"
flake8 = "^6.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[flake8]
max-line-length = 88
